requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
email-validator

# ▶ JWT
python-jose[cryptography]

# ▶ HTTP 요청 (구글 userinfo)
requests

# ▶ 비밀번호 해시
passlib[bcrypt]

# ▶ FormData 받기
python-multipart

# ▶ Firebase 연동
google-auth>=2.0.0
firebase-admin==6.3.0

# ▶ protobuf 버전 명시 (중요!)
protobuf<5.0.0

# ▶ RAG 챗봇용
qdrant-client
openai

transformers
torch
sentence-transformers